Add files via upload
This program takes the users cell number, call length and the number of relay stations and calculates the net cost, the call tax and the total cost of the call. The program then prints the results back to the user. This time the program uses functions to make the calculations, get the input, and print the results.